Попытка реализации сохранения новой точки маршрута. Также реализация безопасности пользовательского ввода
