+++ src/main.js
@@ -6,9 +6,12 @@ import FilterModel from './model/filter-model.js';
 const headerElement = document.querySelector('.page-header__container');
 const pageMainElement = document.querySelector('.page-main');
 const pageMainSortElement = pageMainElement.querySelector('.trip-events');
+const newEventButton = document.querySelector('.trip-main__event-add-btn');
 
 const eventsModel = new EventsModel();
 const filterModel = new FilterModel();
+
+newEventButton.addEventListener('click', handleNewEventButtonClick);
 
 const headerPresenter = new HeaderPresenter({
   headerContainer: headerElement,
@@ -20,7 +23,17 @@ const pagePresenter = new PagePresenter({
   pageContainer: pageMainSortElement,
   eventsModel,
   filterModel,
+  onNewEventDestroy: handleNewEventFormClose
 });
+
+function handleNewEventFormClose() {
+  newEventButton.disabled = false;
+}
+
+function handleNewEventButtonClick() {
+  pagePresenter.createEvent();
+  newEventButton.disabled = true;
+}
 
 headerPresenter.init();
 pagePresenter.init();
